Flattens functions back into single file due to twilio limitations
