--- lib/sentry.js
@@ -1,14 +1,0 @@
-const Sentry = require("@sentry/node");
-
-const Tracing = require("@sentry/tracing");
-
-Sentry.init({
-  dsn: process.env.SENTRY_DSN,
-
-  // Set tracesSampleRate to 1.0 to capture 100%
-  // of transactions for performance monitoring.
-  // We recommend adjusting this value in production
-  tracesSampleRate: process.env.SENTRY_TRACE_SAMPLE_RATE || 0,
-});
-
-module.exports = Sentry;
